requirements.txt
streamlit
google-genai
pypdf
faiss-cpu
numpy
langchain
langchain-google-genai
langchain-text-splitters